+++ package.json
@@ -29,7 +29,6 @@
     "@tailwindcss/typography": "^0.5.3",
     "algoliasearch": "^4.13.1",
     "classnames": "^2.3.1",
-    "core-js-pure": "^3.23.4",
     "framer-motion": "^6.4.3",
     "gatsby": "^4.18.0",
     "gatsby-plugin-algolia-search": "^0.5.7",
