+++ src/utils/get-month-and-day.js
@@ -1,0 +1,6 @@
+export default function getMonthAndDay(date) {
+  const convertedDate = new Date(date);
+  const day = convertedDate.getUTCDate();
+  const month = convertedDate.toLocaleString('default', { month: 'long' });
+  return `${month}, ${day}`;
+}
