+++ src/utils/get-year.js
@@ -1,0 +1,5 @@
+export default function getYear(date) {
+  const convertedDate = new Date(date);
+  const year = convertedDate.getFullYear();
+  return year;
+}
